+++ ubo_app/menu_app/menu.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 from typing_extensions import override
 from ubo_gui.app import UboApp
 
@@ -12,9 +14,18 @@ from ubo_app.store.main import store
 from ubo_app.store.services.display import DisplayRerenderEvent
 from ubo_app.store.settings.types import SettingsSetDebugModeEvent
 
+if TYPE_CHECKING:
+    from ubo_app.utils.types import Subscriptions
+
 
 class MenuApp(MenuAppCentral, MenuAppFooter, MenuAppHeader, UboApp):
     """Menu application."""
+
+    def __init__(self: MenuApp, **kwargs: object) -> None:
+        """Initialize the application."""
+        super().__init__(**kwargs)
+        self.subscriptions: Subscriptions = []
+        self.is_stopped = False
 
     def set_debug_mode(self: MenuApp, event: SettingsSetDebugModeEvent) -> None:
         """Set the debug mode."""
@@ -27,8 +38,6 @@ class MenuApp(MenuAppCentral, MenuAppFooter, MenuAppHeader, UboApp):
     @override
     def on_start(self: MenuApp) -> None:
         """Start the application."""
-        self.is_stopped = False
-
         from ubo_app.side_effects import setup_side_effects
 
         self.subscriptions = setup_side_effects()
@@ -41,8 +50,10 @@ class MenuApp(MenuAppCentral, MenuAppFooter, MenuAppHeader, UboApp):
         store.subscribe_event(DisplayRerenderEvent, self.rerender, keep_ref=False)
 
     @override
-    def on_stop(self: MenuApp) -> None:
+    def stop(self, *largs: object) -> object:
         """Stop the application."""
+        return_value = super().stop(*largs)
+        self.is_stopped = True
         for cleanup in self.subscriptions:
             cleanup()
-        self.is_stopped = True
+        return return_value