+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - refactor(core): use latest version of `python-redux` and `ubo-gui` and remove all `@mainthread` decorators for actions on the menu as `ubo-gui` now takes care of running them in the main thread itself
 - fix(core): improve resource cleanup of service threads so that they can be restarted after being stopped
 - refactor(core): defer attaching the main reducer to the root reducer using `CombineReducerRegisterAction` to avoid circular imports or sacrificing the purity of reducers.
+- fix(core): add cleanup for gpio pins including display pins and run `gpiozero.devices._shutdown` as part of the cleanup process
 
 ## Version 1.2.2